+++ client/src/Components/LabelMilestoneButton.js
@@ -1,0 +1,71 @@
+import React from 'react';
+import styled from 'styled-components';
+import { LabelIcon, MilestoneIcon } from './icons';
+
+const Wrapper = styled.div`
+  display: flex;
+`;
+
+const LabelButton = styled.button`
+  display: flex;
+  align-items: center;
+  height: 60px;
+  padding: 10px 20px;
+  border: 1px solid ${(props) => props.theme.grayColor};
+  border-radius: ${(props) => props.theme.radius} 0 0 ${(props) => props.theme.radius};
+  background-color: ${(props) => props.theme.skyblueColor};
+  svg {
+    fill: ${(props) => props.theme.whiteColor};
+  }
+`;
+
+const MilestoneButton = styled(LabelButton)`
+  border-radius: 0 ${(props) => props.theme.radius} ${(props) => props.theme.radius} 0;
+`;
+
+const ButtonTitle = styled.span`
+  margin-left: 5px;
+`;
+
+const Count = styled.span`
+  display: inline-block;
+  padding: 2px 7px;
+  margin-left: 10px;
+  border-radius: 15px;
+  background-color: ${(props) => props.theme.blueColor};
+  color: ${(props) => props.theme.whiteColor};
+`;
+
+const LabelMilestoneButton = ({ isCount }) => {
+  if (isCount) {
+    return (
+      <Wrapper>
+        <LabelButton>
+          <LabelIcon />
+          <ButtonTitle>Labels</ButtonTitle>
+          <Count>3</Count>
+        </LabelButton>
+        <MilestoneButton>
+          <MilestoneIcon />
+          <ButtonTitle>Milestones</ButtonTitle>
+          <Count>10</Count>
+        </MilestoneButton>
+      </Wrapper>
+    );
+  }
+
+  return (
+    <Wrapper>
+      <LabelButton>
+        <LabelIcon />
+        <ButtonTitle>Labels</ButtonTitle>
+      </LabelButton>
+      <MilestoneButton>
+        <MilestoneIcon />
+        <ButtonTitle>Milestones</ButtonTitle>
+      </MilestoneButton>
+    </Wrapper>
+  );
+};
+
+export default LabelMilestoneButton;
