+++ client/src/theme.js
@@ -35,6 +35,12 @@ export const GlobalStyle = createGlobalStyle`
   }
   button {
       border: none;
+      outline: none;
+      border-radius: ${(props) => props.theme.radius};
+      color: ${(props) => props.theme.whiteColor};
+      font-size: 18px;
+      font-weight: 700;
+      cursor: pointer;
   }
   p {
       margin: 0;
